tests: reduce internet access in tests
