+++ tests/installation/test_executor.py
@@ -11,7 +11,6 @@ from subprocess import CalledProcessError
 from typing import TYPE_CHECKING
 from typing import Any
 from typing import Callable
-from urllib.parse import urlparse
 
 import pytest
 
@@ -40,9 +39,6 @@ from tests.repositories.test_pypi_repository import MockRepository
 if TYPE_CHECKING:
     from collections.abc import Iterator
 
-    import httpretty
-
-    from httpretty.core import HTTPrettyRequest
     from pytest_mock import MockerFixture
 
     from poetry.config.config import Config
@@ -132,36 +128,6 @@ def pool() -> RepositoryPool:
     pool.add_repository(MockRepository())
 
     return pool
-
-
-@pytest.fixture
-def mock_file_downloads(
-    http: type[httpretty.httpretty], fixture_dir: FixtureDirGetter
-) -> None:
-    def callback(
-        request: HTTPrettyRequest, uri: str, headers: dict[str, Any]
-    ) -> list[int | dict[str, Any] | bytes]:
-        name = Path(urlparse(uri).path).name
-
-        fixture = Path(__file__).parent.parent.joinpath(
-            "repositories/fixtures/pypi.org/dists/" + name
-        )
-
-        if not fixture.exists():
-            fixture = fixture_dir("distributions") / name
-
-            if not fixture.exists():
-                fixture = (
-                    fixture_dir("distributions") / "demo-0.1.0-py2.py3-none-any.whl"
-                )
-
-        return [200, headers, fixture.read_bytes()]
-
-    http.register_uri(
-        http.GET,
-        re.compile("^https://files.pythonhosted.org/.*$"),
-        body=callback,
-    )
 
 
 @pytest.fixture
@@ -716,6 +682,7 @@ def test_executor_should_write_pep610_url_references_for_non_wheel_files(
     config: Config,
     io: BufferedIO,
     fixture_dir: FixtureDirGetter,
+    mock_file_downloads: None,
 ) -> None:
     url = (fixture_dir("distributions") / "demo-0.1.0.tar.gz").resolve()
     package = Package("demo", "0.1.0", source_type="file", source_url=url.as_posix())
